scripts/requirements.txt
python-docx
flask
python-jobspy